[UserData_1.0] Template functions to verify username and password
Sever side python script contain functions to read/write to a json file, where the users security keys will be stored.
Future plans is to implement file signing, to add an extra layer of securtiy to these machines. Thus, a wrapper of reading the
json files is included but not filled out.
This should not be merged until a HTTP server which can receive messages as headers and send server pushes to devices is built.